Remove the elastic registry and just keep 4 registries as proxies to the Docker Hub.
Previously, we created a specific registry which was connecting to the Elastic's docker registry but the problem is that its connection was not reliable and at sometimes it asked for authentication. So we removed this registry altogether and instead decided to use only our main registry for these kind of images that could not be found on Docker Hub. We simply pull these images from wherever they, tag them by prepending our registry address to it and then push them inside our own registry to be usable by other servers.

We still keep those 4 registries that act as a proxy to Docker Hub.